fix: resolve remaining lint errors
- Extract useProjectStats to features/projects/hooks/use-project-stats.ts
- Add eslint-disable for badgeVariants co-export in badge.tsx (shadcn pattern)
- Fix circular useCallback deps in task-provider using refs for undo actions
