+++ src/shared/components/ui/badge.tsx
@@ -43,4 +43,5 @@ function Badge({
   );
 }
 
+// eslint-disable-next-line react-refresh/only-export-components
 export { Badge, badgeVariants };
